backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
playwright
beautifulsoup4
anthropic
apscheduler
thefuzz
